Added a `Bootloader` 💿
It's time to test launch our simple kernel for the first time! 
We are currently using an external bootloader, but we will create our own soon, but let us focus on the kernel first.
To start the current version of our OS (Assuming that the env is already set up):
- First: we need to create a build of our kernel and link the generated binary into our bootloader: `cargo bootimage`.
- We can use the `.bin` image generated in: `target/x86-64-crab_os/debug/bootimage-crab-os.bin`.
- We can run it using an external USB drive, however since  we are still testing, it's better to use it with an emulator.
- To run with `qemu`: `qemu-system-x86_64 -drive format=raw,file=target/x86-64-crab_os/debug/bootimage-crab-os.bin`.