Added `VGA Buffer` handler 🖍️
We now have a basic IO handling. we can now print text to the screen without directly modifying the VGA buffer in memory, we now have a handler that gets most of the basic work for us. And we have redefined the macros (print/println) that we have lost since we are not using the std provided by rust!